Reuse core game logic and add features
