+++ cinco/src/main/kotlin/ca/allanwang/discord/bot/cinco/CincoBot.kt
@@ -2,10 +2,11 @@ package ca.allanwang.discord.bot.cinco
 
 
 import ca.allanwang.discord.bot.base.*
-import ca.allanwang.discord.bot.cinco.game.CincoContext
 import ca.allanwang.discord.bot.cinco.game.CincoGame
-import ca.allanwang.discord.bot.cinco.game.CincoGameModule
-import ca.allanwang.discord.bot.cinco.game.CincoVariant
+import ca.allanwang.discord.bot.cinco.game.WordBank
+import ca.allanwang.discord.bot.cinco.game.CincoContext
+import ca.allanwang.discord.bot.cinco.game.features.CincoGameFeatureModule
+import ca.allanwang.discord.bot.cinco.game.features.CincoVariant
 import com.gitlab.kordlib.core.behavior.channel.MessageChannelBehavior
 import com.gitlab.kordlib.core.behavior.channel.createEmbed
 import com.gitlab.kordlib.core.behavior.edit
@@ -25,7 +26,8 @@ import javax.inject.*
 
 @Singleton
 class CincoBot @Inject constructor(
-    private val cincoProvider: Provider<CincoComponent.Builder>
+    private val cincoProvider: Provider<CincoComponent.Builder>,
+    private val wordBank: WordBank
 ) : CommandHandlerBot {
 
     companion object {
@@ -46,7 +48,24 @@ class CincoBot @Inject constructor(
                     }
                 }
             }
+            arg("check") {
+                action(withMessage = true) {
+                    checkWord()
+                }
+            }
         }
+    }
+
+    private suspend fun CommandHandlerEvent.checkWord() {
+        val word = message.trim()
+        val isWord = wordBank.isWord(word)
+        channel.createMessage(buildString {
+            appendCodeBlock {
+                append(word)
+            }
+            if (isWord) append(" is a registered word")
+            else append(" is not a registered word")
+        })
     }
 
     private suspend fun CommandHandlerEvent.selectVariant() {
@@ -104,7 +123,7 @@ class CincoBot @Inject constructor(
                 CincoContext(
                     botPrefix = prefix,
                     gameRounds = 15,
-                    roundTimeout = 15_000L
+                    roundTimeout = 30_000L
                 )
             )
             .build()
@@ -124,7 +143,7 @@ annotation class CincoPlayers
 interface CincoModule
 
 @CincoScope
-@Subcomponent(modules = [CincoGameModule::class])
+@Subcomponent(modules = [CincoGameFeatureModule::class])
 interface CincoComponent {
 
     @CincoScope
