+++ cinco/src/main/kotlin/ca/allanwang/discord/bot/cinco/game/CincoContext.kt
@@ -1,0 +1,7 @@
+package ca.allanwang.discord.bot.cinco.game
+
+data class CincoContext(
+    val botPrefix: String,
+    val gameRounds: Int,
+    val roundTimeout: Long
+)
