--- cinco/src/main/kotlin/ca/allanwang/discord/bot/cinco/game/CincoGameModule.kt
@@ -1,14 +1,0 @@
-package ca.allanwang.discord.bot.cinco.game
-
-import ca.allanwang.discord.bot.cinco.CincoScope
-import dagger.Module
-import dagger.Provides
-
-@Module
-object CincoGameModule {
-    @Provides
-    @CincoScope
-    fun cinco(variant: CincoVariant, cincoAzul: CincoAzul): CincoGame = when (variant) {
-        CincoVariant.Azul -> cincoAzul
-    }
-}
