+++ cinco/src/main/kotlin/ca/allanwang/discord/bot/cinco/game/features/CincoGameFeature.kt
@@ -1,0 +1,17 @@
+package ca.allanwang.discord.bot.cinco.game.features
+
+import ca.allanwang.discord.bot.cinco.game.CincoContext
+import com.gitlab.kordlib.rest.builder.message.EmbedBuilder
+
+interface CincoGameFeature {
+    val cincoContext: CincoContext
+    suspend fun startRound(round: Int)
+    suspend fun skip(round: Int)
+    suspend fun endGame()
+    fun EmbedBuilder.roundProgressFooter(round: Int) {
+        footer {
+            text = "$round/${cincoContext.gameRounds}"
+        }
+    }
+}
+
